+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 0.2.0
 ### Friday, November 20, 2015
 
+- (Experimental) Added a new "Run selection" (F8) command which executes the current code selection and displays the output
 - Added a new online help command!  Press Ctrl+F1 to get help for the symbol under the cursor.
 - Enabled PowerShell language features for untitled and in-memory (e.g. in Git diff viewer) PowerShell files
 - Added `powershell.scriptAnalysis.enable` configuration variable to allow disabling script analysis for performance (issue #11)